+++ packages/agtree/CHANGELOG.md
@@ -8,6 +8,14 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
+## [3.2.0] - 2025-05-22
+
+### Added
+
+- Support for converting HTML rules to uBlock syntax.
+
+[3.2.0]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/agtree-3.2.0
+
 ## [3.1.5] - 2025-05-20
 
 ### Added