+++ packages/tswebextension/CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.0-alpha.3] - 2025-05-22
+
+### Fixed
+
+- Improved rule set caching to avoid inconsistent behaviors.
+- Improved rule set update in declarative filtering log to avoid unexpected behaviors during configuration changes.
+
+[3.1.0-alpha.3]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tswebextension-v3.1.0-alpha.3
+
 ## [3.1.0-alpha.2] - 2025-05-19
 
 ### Changed